pyproject.toml
[tool.poetry]
name = "seekout"
version = "0.1.2"
description = "Library for parsing web pages"
authors = ["mauza <mauza@mauza.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
requests = "^2.24.0"
beautifulsoup4 = "^4.9.3"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
flake8 = "^3.8.4"
black = "^20.8b1"
coverage = "^5.3"
pre-commit = "^2.8.2"
pre-commit-hooks = "^3.3.0"
safety = "^1.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
